Draft: Resolve "Code has warnings when included in Autoware"

Closes #1

See merge request ApexAI/point_cloud_msg_wrapper!1

+++ point_cloud_msg_wrapper/include/point_cloud_msg_wrapper/field_generators.hpp
@@ -50,7 +50,7 @@ void push_back_field(
   fields.emplace_back();
   fields.back()
   .set__name(name)
-  .set__offset(offset_of(member))
+  .set__offset(static_cast<std::uint32_t>(offset_of(member)))
   .set__datatype(get_field_datatype<MemberT>())
   .set__count(get_field_count<MemberT>());
 }
